refactor(oauth): export the names that have consumers
The narrow __all__ needed three paragraphs of docstring to explain why the
login flows were reachable but unadvertised — an apology for a rule the code
did not enforce, since bfabric_scripts (a separate distribution) imported five
of those submodules anyway. A name imported across a distribution boundary is
public in fact; the surface was drawn in the wrong place.

Export the nine names that have a consumer, so nothing has to reach into a
submodule for a supported name, and the docstring is left with only what is
actually non-obvious: that the package root pulls authlib and joserfc while
`import bfabric` does not. The contributor-facing rule — underscore-prefix a
module exactly when nothing outside the package imports it, hence only
_token_exchange — moves to the design doc, where it costs nothing per session.

bfabric.py keeps its function-local submodule imports: routing them through
the root would drag joserfc and authlib into `import bfabric`.

Measured on the real CLI path (bfabric_scripts.cli.login.manage), the wider
root costs ~1 ms — those deps were already loaded there.
